requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pytest==8.3.5
pytest-asyncio==0.26.0
azure-identity==1.19.0
